* 修复日志无法删除问题 * 数据库   * 重构 mysql profiles信息查看，改为弹窗   * 数据库查询的数据可以通过右击导出     * 可以导出 SQL、Excel、CSV、Txt   * 数据导出 Excel、CSV、Txt 第一行导出列名

+++ internal/module/module_database/api.go
@@ -956,10 +956,10 @@ func (this_ *api) taskClean(requestBean *base.RequestBean, c *gin.Context) (res 
 	task := worker.GetTask(request.TaskId)
 	if task != nil {
 		if task.Extend != nil {
-			if task.Extend["dirPath"] != "" {
+			if task.Extend["dirPath"] != nil && task.Extend["dirPath"] != "" {
 				_ = os.RemoveAll(task.Extend["dirPath"].(string))
 			}
-			if task.Extend["zipPath"] != "" {
+			if task.Extend["zipPath"] != nil && task.Extend["zipPath"] != "" {
 				_ = os.Remove(task.Extend["zipPath"].(string))
 			}
 		}
